Bugfix alias_map_file not being used for table aliases (#1492)
* Add failing test for using configured alias map

The test demonstrates a bug in the PGCompleter class where a configured
alias map file is not used to generate aliases.

* Use configured alias map in PGCompleter

Fixes a bug where this configuration was ignored.

* Update generate_alias docstring

Clarify the behaviour of generate_alias when an alias_map argument is
passed.

* Add "no match in map" test case for generate_alias

Test demonstrates passing an alias_map to generate_alias does not break
generation of aliases for tables not covered within the map file.

* Update AUTHORS file

Credit where credit is due ;)