Add suggestion when setting the search_path (#1513)
To be suggested autocompletion when setting the search_path, the
keyword's suggestion "search_path TO" will be suggested on a "SET"
token, only if the latter is the first token of the SQL statement (to
differentiate it from the "UPDATE foo SET"). Then, available schemas are
suggested (after the "TO" token).

Resolves #1511